Refactor: Use Start-Process
Note: Start-Process requires that each item in the -ArgumentList must be double-quoted if there are spaces, or else they be treated as individual arguments. See: https://github.com/PowerShell/PowerShell/issues/5576

Hence it does becomes a bit clumsy to build `-ArgumentList`. And it means that if each item's value has spaces and also has a double quote, the double quote must be escaped to be wrapped by outer double-quotes.